adjustable shaft toleraance
An additional parameter "hub_squeezeout_factor" allows to adjust the
tightness of fit.
Normally the center of the extusion path is moved by 1/2 of the
extrusion width so that the surface is at the desired position, but
because the extrusion paths are crossing each other, the effective
extrusion width is wider.